postNotes/pyproject.toml
[tool.poetry]
name = "postnotes"
version = "0.1.0"
description = ""
authors = ["Amontears <amontearx66@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.10"
django = "^5.1.3"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
